requirements.txt
numpy
scikit-image
scikit-learn
scipy
matplotlib
h5py
pandas
opencv-python
tensorflow-hub
jsmin
opendatalake

# Dev/Deployment
#sphinx
#sphinx_rtd_theme
#nose
#coverage
#pypi-publisher
